requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# OpenClaw Gateway
websockets>=12.0
httpx>=0.26.0

# RAG (复用 claw_with_rag 的依赖)
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.0
pyyaml>=6.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0